Implemented Repository Pattern for the User GetAsync method.
Created the IRepository.cs interface which defines the functions.

Created the Repository.cs implementation class that inherits IRepository Interface and implemented functionality for the GetAll.

Replaced the method in the Get Endpoint of the API to use the repository implementation of GetAsync().

Added the dependency injection of the repository to the Program.cs file.