Completed Jwt Authentication mechanism. Now we can use the /auth/login endpoint and if the username and password are correct, you'll receive a Jwt Token with 30 minutes Expiration date.
For this, I created an interface for authentication service IAuthenticationService, and used it in the AuthenticationEndpoint following the repository pattern.

I modified the Post method in the User endpoints to use AspNetCore.Identity package instead o BCrypt, so then on the authentication endpoint I could rely on the VerifyPassword method to compared Hashed Passwords and complete the authentication mechanism.

Added the Authentication Endpoint as a dependency to complete the dependency injection of the endpoint.