Implemented in IShoppingCartService and ShoppingCartService the method GetShoppingCartItemByIdAsync() to be able to fetch specific items from the cart.
Modified the PUT Endpoint for shopping cart and the DELETE endpoint that deletes specific items to modify also the total values of the cart.

Modified the DELETE Endpoint that deletes specific products by ID to take into account into the situation where the last item in the cart is deleted this way.