Created the IReportingService interface and ReportingService classes for implementation using Repository Pattern to separate database operations from the endpoints.
Created the ReportsEndpoint.cs file to build the API endpoints to retrieve the data for these reports.

Created the ReportTopCustomersModel to establish a standardized way to group the top customers and then send the details in the request.

Created the ReportTopProductsModel to establish a standardized way to group to products and send then through the API requiest.

Injected in the Program.cs file the IReportingService and the ReportEndpoints.

Added the new endpoints details to the README file.